+++ src/pages/CreatePostPage/hooks/index.ts
@@ -1,0 +1,1 @@
+export {default as useCreatePost} from "./useCreatePost"
