+++ src/pages/CreatePostPage/hooks/useCreatePost.ts
@@ -1,0 +1,96 @@
+import { ChangeEvent, FocusEvent, useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useFetchCreatePost } from '@apis/post';
+import { joinDataBySeparator } from '@utils/parseDataBySeparator';
+
+const MAX_TITLE_OPTION_LENGTH = 100;
+
+const useCreatePost = () => {
+  const navigate = useNavigate();
+  const {
+    createPostMutate,
+    createPostData,
+    isCreatePostLoading,
+    isCreatePostSuccess,
+    isCreatePostError,
+  } = useFetchCreatePost();
+
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const [inputValues, setInputValues] = useState({
+    title: '',
+    optionA: '',
+    optionB: '',
+  });
+
+  const isCreatePostPossible: boolean =
+    inputValues.title.length > 0 &&
+    inputValues.optionA.length > 0 &&
+    inputValues.optionB.length > 0;
+
+  const handleChangeTitleValue = (e: ChangeEvent<HTMLInputElement>) => {
+    const { value } = e.target;
+    setInputValues({
+      ...inputValues,
+      title: value.substring(0, MAX_TITLE_OPTION_LENGTH),
+    });
+  };
+
+  const handleChangeOptionValues = (e: ChangeEvent<HTMLTextAreaElement>) => {
+    const { id, value } = e.target;
+    if (e.target.scrollHeight === e.target.clientHeight) {
+      setInputValues({
+        ...inputValues,
+        [id]: value.substring(0, MAX_TITLE_OPTION_LENGTH),
+      });
+    }
+  };
+
+  const handleBlurTrim = (
+    e: FocusEvent<HTMLInputElement | HTMLTextAreaElement>,
+  ) => {
+    const { id, value } = e.target;
+    setInputValues({
+      ...inputValues,
+      [id]: value.trim(),
+    });
+  };
+
+  const handleClickCreatePost = () => {
+    if (!isCreatePostPossible) return;
+
+    const postTitle = joinDataBySeparator(
+      inputValues.title,
+      inputValues.optionA,
+      inputValues.optionB,
+    );
+
+    createPostMutate({ title: postTitle });
+  };
+
+  useEffect(() => {
+    if (isCreatePostSuccess) {
+      navigate(`/post/${createPostData}`);
+    }
+  }, [isCreatePostSuccess]);
+
+  useEffect(() => {
+    if (isCreatePostError) {
+      setIsModalOpen(() => true);
+    }
+  }, [isCreatePostError]);
+
+  return {
+    inputValues,
+    isModalOpen,
+    setIsModalOpen,
+    isCreatePostLoading,
+    isCreatePostPossible,
+    handleChangeTitleValue,
+    handleChangeOptionValues,
+    handleClickCreatePost,
+    handleBlurTrim,
+  };
+};
+
+export default useCreatePost;
